fix(xrpl): package import renaming
