+++ xrpl/client/websocket/ledger.go
@@ -1,8 +1,8 @@
 package websocket
 
 import (
-	"github.com/Peersyst/xrpl-go/xrpl/model/client/common"
-	"github.com/Peersyst/xrpl-go/xrpl/model/client/ledger"
+	"github.com/Peersyst/xrpl-go/xrpl/model/requests/common"
+	"github.com/Peersyst/xrpl-go/xrpl/model/requests/ledger"
 )
 
 const (
